+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxAuxData.hs
@@ -1,9 +1,11 @@
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Conway.TxAuxData () where
 
 import Cardano.Ledger.Alonzo.Core
+import Cardano.Ledger.Alonzo.Plutus.Context (SupportedPlutusRunnable (..))
 import Cardano.Ledger.Alonzo.TxAuxData (
   AlonzoTxAuxData (..),
   metadataAlonzoTxAuxDataL,
@@ -13,8 +15,12 @@ import Cardano.Ledger.Alonzo.TxAuxData (
  )
 import Cardano.Ledger.Conway.Era
 import Cardano.Ledger.Conway.Scripts ()
+import Data.Map.Strict (Map)
 
-instance EraTxAuxData ConwayEra where
+instance
+  StAnnTxCache ConwayEra ~ Map ScriptHash (SupportedPlutusRunnable ConwayEra) =>
+  EraTxAuxData ConwayEra
+  where
   type TxAuxData ConwayEra = AlonzoTxAuxData ConwayEra
 
   mkBasicTxAuxData = AlonzoTxAuxData mempty mempty mempty
@@ -23,8 +29,14 @@ instance EraTxAuxData ConwayEra where
 
   validateTxAuxData = validateAlonzoTxAuxData
 
-instance AllegraEraTxAuxData ConwayEra where
+instance
+  StAnnTxCache ConwayEra ~ Map ScriptHash (SupportedPlutusRunnable ConwayEra) =>
+  AllegraEraTxAuxData ConwayEra
+  where
   nativeScriptsTxAuxDataL = nativeScriptsAlonzoTxAuxDataL
 
-instance AlonzoEraTxAuxData ConwayEra where
+instance
+  StAnnTxCache ConwayEra ~ Map ScriptHash (SupportedPlutusRunnable ConwayEra) =>
+  AlonzoEraTxAuxData ConwayEra
+  where
   plutusScriptsTxAuxDataL = plutusScriptsAllegraTxAuxDataL