+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/TxAuxData.hs
@@ -1,8 +1,11 @@
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Dijkstra.TxAuxData () where
 
+import Cardano.Ledger.Alonzo.Core
+import Cardano.Ledger.Alonzo.Plutus.Context (SupportedPlutusRunnable (..))
 import Cardano.Ledger.Alonzo.TxAuxData (
   AlonzoTxAuxData (..),
   metadataAlonzoTxAuxDataL,
@@ -10,15 +13,14 @@ import Cardano.Ledger.Alonzo.TxAuxData (
   plutusScriptsAllegraTxAuxDataL,
   validateAlonzoTxAuxData,
  )
-import Cardano.Ledger.Conway.Core (
-  AllegraEraTxAuxData (..),
-  AlonzoEraTxAuxData (..),
-  EraTxAuxData (..),
- )
 import Cardano.Ledger.Dijkstra.Era (DijkstraEra)
 import Cardano.Ledger.Dijkstra.Scripts ()
+import Data.Map.Strict (Map)
 
-instance EraTxAuxData DijkstraEra where
+instance
+  StAnnTxCache DijkstraEra ~ Map ScriptHash (SupportedPlutusRunnable DijkstraEra) =>
+  EraTxAuxData DijkstraEra
+  where
   type TxAuxData DijkstraEra = AlonzoTxAuxData DijkstraEra
 
   mkBasicTxAuxData = AlonzoTxAuxData mempty mempty mempty
@@ -27,8 +29,14 @@ instance EraTxAuxData DijkstraEra where
 
   validateTxAuxData = validateAlonzoTxAuxData
 
-instance AllegraEraTxAuxData DijkstraEra where
+instance
+  StAnnTxCache DijkstraEra ~ Map ScriptHash (SupportedPlutusRunnable DijkstraEra) =>
+  AllegraEraTxAuxData DijkstraEra
+  where
   nativeScriptsTxAuxDataL = nativeScriptsAlonzoTxAuxDataL
 
-instance AlonzoEraTxAuxData DijkstraEra where
+instance
+  StAnnTxCache DijkstraEra ~ Map ScriptHash (SupportedPlutusRunnable DijkstraEra) =>
+  AlonzoEraTxAuxData DijkstraEra
+  where
   plutusScriptsTxAuxDataL = plutusScriptsAllegraTxAuxDataL